Account for adding/removing weight by discarding input
Because the scale is continuously sending input, we need to handle the
case where weight is being add/removed from the scale and not treat those
values as legitimate weight data. The fix for this is to discard the input
for a set amount of time, then check that the values read before and after
the junk lines are the same.